Do not disable coloring when arguments contains -l
Arguments could be misinterpreted and disable coloring when not needed.

For example:

    mvn dependency:tree -Dincludes="commons-logging"

Instead of just testing if -l, --log-file, -B, --batch-mode are contained
in MAVEN_CMD_LINE_ARGS, arguments are now split on space and equals sign.
Color are disabled if one of the fragment equals log file or batch mode option.

Fixes #26