Aggregate JaCoCo coverage reports
By default JaCoCo reports are specific to a module.

For example LogLevelActivation (in maven-color-core) was reported as 0% code coverage
because there is not UT for this class in the same module. But this code is tested by
a UT in maven-color-gossip.

There is a report-aggregate in jacoco-maven-plugin to remediate this.
This is not really well documented but there is an example in: https://github.com/jacoco/jacoco/tree/master/jacoco-maven-plugin.test/it/it-report-aggregate